+++ src/idc/tool/convert.py
@@ -10,7 +10,7 @@ from seppl.io import execute, Reader, Filter, MultiFilter, Writer
 from wai.logging import init_logging, set_logging_level, add_logging_level, LOGGING_LEVELS
 from idc.core import Session, ENV_IDC_LOGLEVEL
 from idc.help import generate_plugin_usage
-from idc.registry import available_readers, available_filters, available_writers
+from idc.registry import available_readers, available_filters, available_writers, REGISTRY
 
 
 CONVERT = "idc-convert"
@@ -52,9 +52,9 @@ def _print_usage(plugin_details: bool = False):
     print()
     print("Tool for converting between image annotation dataset formats.")
     print()
-    print("readers (%d):\n" % len(available_readers()) + enumerate_plugins(available_readers().keys(), prefix="   "))
-    print("filters (%d):\n" % len(available_filters()) + enumerate_plugins(available_filters().keys(), prefix="   "))
-    print("writers (%d):\n" % len(available_writers()) + enumerate_plugins(available_writers().keys(), prefix="   "))
+    print("readers (%d):\n" % len(available_readers()) + enumerate_plugins(available_readers().keys(), aliases=REGISTRY.all_aliases, prefix="   "))
+    print("filters (%d):\n" % len(available_filters()) + enumerate_plugins(available_filters().keys(), aliases=REGISTRY.all_aliases, prefix="   "))
+    print("writers (%d):\n" % len(available_writers()) + enumerate_plugins(available_writers().keys(), aliases=REGISTRY.all_aliases, prefix="   "))
     print()
     print("optional arguments:")
     print("  -h, --help            show basic help message and exit")